289: Refactor source folder
