+++ src/layers/index.ts
@@ -2,11 +2,11 @@ export * from './Field';
 export * from './LayerGroup';
 export * from './Projection';
 export * from './FeatureLayer';
-export * from './FeatureSource';
+export * from './sources/FeatureSource';
 export * from './PropertyAggregator';
-export * from './MemoryFeatureSource';
-export * from './GeoJSONFeatureSource';
-export * from './ShapefileFeatureSource';
-export * from './FeatureSourceFactory';
+export * from './sources/MemoryFeatureSource';
+export * from './sources/GeoJSONFeatureSource';
+export * from './sources/ShapefileFeatureSource';
+export * from './sources/FeatureSourceFactory';
 export * from './LayerFactory';
 export * from 'ginkgoch-shapefile';
